generic design for scene benches
